# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VoiceCategory = (typeof VoiceCategory)[keyof typeof VoiceCategory];

export const VoiceCategory = {
  free: "free",
  premium: "premium",
  cloned: "cloned",
} as const;

export type VoiceGender = (typeof VoiceGender)[keyof typeof VoiceGender];

export const VoiceGender = {
  male: "male",
  female: "female",
  neutral: "neutral",
} as const;

export interface Voice {
  id: string;
  name: string;
  category: VoiceCategory;
  gender: VoiceGender;
  accent: string;
  language: string;
  previewUrl?: string;
  description: string;
  tags: string[];
  isCloned: boolean;
}

export interface TtsRequest {
  text: string;
  voiceId: string;
  stability?: number;
  similarity?: number;
  style?: number;
}

export interface TtsResponse {
  jobId: string;
  status: string;
  creditsUsed: number;
  durationSeconds: number;
  audioUrl?: string;
  text: string;
  voiceId: string;
  createdAt: string;
}

export interface CloneRequest {
  name: string;
  description?: string;
  audioBase64: string;
  fileName: string;
}

export interface CloneResponse {
  voiceId: string;
  name: string;
  status: string;
  creditsUsed: number;
  message: string;
}

export interface HistoryItem {
  id: string;
  text: string;
  voiceId: string;
  voiceName: string;
  durationSeconds: number;
  creditsUsed: number;
  audioUrl?: string;
  createdAt: string;
  status: string;
}

export interface PricingPlan {
  id: string;
  name: string;
  priceNaira: number;
  credits: number;
  features: string[];
  isFree: boolean;
  isPopular: boolean;
}

export interface CreditPack {
  id: string;
  credits: number;
  priceNaira: number;
  label: string;
  durationLabel: string;
}

export interface CreditBalance {
  credits: number;
  plan: string;
  renewsAt?: string;
  totalUsed: number;
}

export interface VoiceStats {
  totalVoices: number;
  totalGenerations: number;
  totalMinutes: number;
  clonedVoices: number;
  activeUsers: number;
}

export type ListVoicesParams = {
  category?: ListVoicesCategory;
};

export type ListVoicesCategory =
  (typeof ListVoicesCategory)[keyof typeof ListVoicesCategory];

export const ListVoicesCategory = {
  all: "all",
  free: "free",
  premium: "premium",
  cloned: "cloned",
} as const;

export type ListVoices200 = {
  voices: Voice[];
};

export type GetHistoryParams = {
  limit?: number;
};

export type GetHistory200 = {
  history: HistoryItem[];
};

export type GetPricing200 = {
  plans: PricingPlan[];
  creditPacks: CreditPack[];
};