# lib/api-zod/src/generated/types/cloneRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CloneRequest {
  name: string;
  description?: string;
  audioBase64: string;
  fileName: string;
}