# lib/api-zod/src/generated/types/cloneResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CloneResponse {
  voiceId: string;
  name: string;
  status: string;
  creditsUsed: number;
  message: string;
}