# lib/api-zod/src/generated/types/creditBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreditBalance {
  credits: number;
  plan: string;
  renewsAt?: string;
  totalUsed: number;
}