# lib/api-zod/src/generated/types/creditPack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreditPack {
  id: string;
  credits: number;
  priceNaira: number;
  label: string;
  durationLabel: string;
}