# lib/api-zod/src/generated/types/historyItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryItem {
  id: string;
  text: string;
  voiceId: string;
  voiceName: string;
  durationSeconds: number;
  creditsUsed: number;
  audioUrl?: string;
  createdAt: string;
  status: string;
}