# lib/api-zod/src/generated/types/pricingPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface PricingPlan {
  id: string;
  name: string;
  priceNaira: number;
  credits: number;
  features: string[];
  isFree: boolean;
  isPopular: boolean;
}