# lib/api-zod/src/generated/types/ttsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface TtsRequest {
  text: string;
  voiceId: string;
  stability?: number;
  similarity?: number;
  style?: number;
}