# lib/api-zod/src/generated/types/ttsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface TtsResponse {
  jobId: string;
  status: string;
  creditsUsed: number;
  durationSeconds: number;
  audioUrl?: string;
  text: string;
  voiceId: string;
  createdAt: string;
}