# lib/api-zod/src/generated/types/voice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceCategory } from "./voiceCategory";
import type { VoiceGender } from "./voiceGender";

export interface Voice {
  id: string;
  name: string;
  category: VoiceCategory;
  gender: VoiceGender;
  accent: string;
  language: string;
  previewUrl?: string;
  description: string;
  tags: string[];
  isCloned: boolean;
}