# lib/api-zod/src/generated/types/voiceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BMsureplug Voice Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceStats {
  totalVoices: number;
  totalGenerations: number;
  totalMinutes: number;
  clonedVoices: number;
  activeUsers: number;
}